services/mt5-connector/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
MetaTrader5>=5.0.4682
python-dotenv>=1.0.0
aiohttp>=3.9.0

